LPS-96692 Move HypersonicServerClassTestRule to the module since it's only used by CounterLocalServiceTest
