+++ modules/apps/counter/counter-test/bnd.bnd
@@ -1,3 +1,4 @@
 Bundle-Name: Liferay Counter Test
 Bundle-SymbolicName: com.liferay.counter.test
 Bundle-Version: 1.0.0
+-includeresource: @hsqldb-[0-9]*.jar
