+++ modules/apps/counter/counter-test/build.gradle
@@ -1,7 +1,10 @@
 dependencies {
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	testIntegrationCompile group: "org.hsqldb", name: "hsqldb", version: "2.3.3"
 	testIntegrationCompile project(":core:petra:petra-process")
+	testIntegrationCompile project(":core:petra:petra-reflect")
+	testIntegrationCompile project(":core:petra:petra-string")
 	testIntegrationCompile project(":core:registry-api")
 	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
